Setting up packages and refactoring security classes names
