+++ src/main/java/com/github/pedroluiznogueira/securityapi/security/manager/AuthManager.java
@@ -1,24 +1,20 @@
-package com.github.pedroluiznogueira.securityapi.security;
+package com.github.pedroluiznogueira.securityapi.security.manager;
 
+import com.github.pedroluiznogueira.securityapi.security.provider.AuthProvider;
+import com.github.pedroluiznogueira.securityapi.security.filter.AuthFilter;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 
 // this class has 3 important methods
 @Configuration
-public class MyAuthenticationManager extends WebSecurityConfigurerAdapter {
+public class AuthManager extends WebSecurityConfigurerAdapter {
 
     @Autowired
-    private MyAuthenticationProvider authenticationProvider;
+    private AuthProvider authenticationProvider;
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
@@ -31,6 +27,6 @@ public class MyAuthenticationManager extends WebSecurityConfigurerAdapter {
         http.httpBasic();
         http.authorizeRequests().antMatchers("/some/hello").authenticated();
         // best practice
-        http.addFilterBefore(new MyAuthenticationFilter(), BasicAuthenticationFilter.class);
+        http.addFilterBefore(new AuthFilter(), BasicAuthenticationFilter.class);
     }
 }
